rental_application-main/requirement.txt
africastalking==1.2.4
amqp==5.1.1
APScheduler==3.10.0
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contextlib2==21.6.0
Django==4.1.6
django-bootstrap-v5==1.0.11
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
django-multiselectfield==0.1.12
django-rest-authtoken==2.1.4
django-storages==1.13.2
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==3.4
importlib-metadata==6.0.0
jsonfield==3.1.0
kombu==5.2.4
Markdown==3.4.1
mypy-extensions==1.0.0
numpy==1.24.2
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
redis==4.5.1
requests==2.28.2
schedule==1.1.0
schema==0.7.5
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
twilio==7.16.3
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.3.0
zipp==3.13.0
